MTC-145: infrastructure for deals and trades

+++ src/MarginTrading.TradingHistory.SqlRepositories/PositionsHistorySqlRepository.cs
@@ -8,6 +8,7 @@ using Common.Log;
 using Dapper;
 using MarginTrading.TradingHistory.Core.Domain;
 using MarginTrading.TradingHistory.Core.Repositories;
+using MarginTrading.TradingHistory.SqlRepositories.Entities;
 
 namespace MarginTrading.TradingHistory.SqlRepositories
 {
@@ -16,7 +17,7 @@ namespace MarginTrading.TradingHistory.SqlRepositories
         private const string TableName = "PositionsHistory";
 
         private const string CreateTableScript = "CREATE TABLE [{0}](" +
-                                                 @"[OID] [int] NOT NULL IDENTITY (1,1) PRIMARY KEY,
+                                                 @"[OID] [bigint] NOT NULL IDENTITY (1,1) PRIMARY KEY,
 [Id] [nvarchar](64) NOT NULL
 [DealId] [nvarchar](128) NULL,
 [Code] [bigint] NULL,
